+++ command/cp.go
@@ -213,7 +213,7 @@ func NewCopyCommandFlags() []cli.Flag {
 }
 
 func NewCopyCommand() *cli.Command {
-	return &cli.Command{
+	cmd := &cli.Command{
 		Name:               "cp",
 		HelpName:           "cp",
 		Usage:              "copy objects",
@@ -233,6 +233,9 @@ func NewCopyCommand() *cli.Command {
 			return NewCopy(c, false).Run(c.Context)
 		},
 	}
+
+	cmd.BashComplete = getBashCompleteFn(cmd, false, false)
+	return cmd
 }
 
 // Copy holds copy operation flags and states.