+++ e2e/app_test.go
@@ -2,10 +2,13 @@ package e2e
 
 import (
 	"fmt"
+	"net/url"
 	"os"
 	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/peak/s5cmd/command"
 
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/icmd"
@@ -76,7 +79,6 @@ func TestAppDashStat(t *testing.T) {
 	t.Parallel()
 
 	const (
-		bucket                  = "bucket"
 		fileContent             = "this is a file content"
 		src                     = "file1.txt"
 		expectedOutputIfPrinted = "Operation\tTotal\tError\tSuccess\t"
@@ -87,19 +89,19 @@ func TestAppDashStat(t *testing.T) {
 		isPrintExpected bool
 	}{
 		{
-			command:         fmt.Sprintf("--stat --log %v cp s3://bucket/%v .", "trace", src),
+			command:         fmt.Sprintf("--stat --log %v ls", "trace"),
 			isPrintExpected: true,
 		},
 		{
-			command:         fmt.Sprintf("--stat --log %v cp s3://bucket/%v .", "debug", src),
+			command:         fmt.Sprintf("--stat --log %v ls", "debug"),
 			isPrintExpected: true,
 		},
 		{
-			command:         fmt.Sprintf("--stat --log %v cp s3://bucket/%v .", "info", src),
+			command:         fmt.Sprintf("--stat --log %v ls", "info"),
 			isPrintExpected: true,
 		},
 		{
-			command:         fmt.Sprintf("--stat --log %v cp s3://bucket/%v .", "error", src),
+			command:         fmt.Sprintf("--stat --log %v ls", "error"),
 			isPrintExpected: true,
 		},
 		// if level is an empty string, it ignores log levels and uses default.
@@ -118,8 +120,11 @@ func TestAppDashStat(t *testing.T) {
 			t.Parallel()
 			s3client, s5cmd := setup(t)
 
+			bucket := s3BucketFromTestName(t)
+
 			createBucket(t, s3client, bucket)
 			putFile(t, s3client, bucket, src, fileContent)
+
 			cmd := s5cmd(strings.Fields(tc.command)...)
 
 			result := icmd.RunCmd(cmd)
@@ -132,7 +137,6 @@ func TestAppDashStat(t *testing.T) {
 }
 
 func TestAppProxy(t *testing.T) {
-
 	testcases := []struct {
 		name string
 		flag string
@@ -153,6 +157,21 @@ func TestAppProxy(t *testing.T) {
 
 			proxy := httpProxy{}
 			pxyUrl := setupProxy(t, &proxy)
+
+			// set endpoint scheme to 'http'
+			if os.Getenv(s5cmdTestEndpointEnv) != "" {
+				origEndpoint := os.Getenv(s5cmdTestEndpointEnv)
+				endpoint, err := url.Parse(origEndpoint)
+				if err != nil {
+					t.Fatal(err)
+				}
+				endpoint.Scheme = "http"
+				os.Setenv(s5cmdTestEndpointEnv, endpoint.String())
+
+				defer func() {
+					os.Setenv(s5cmdTestEndpointEnv, origEndpoint)
+				}()
+			}
 
 			os.Setenv("http_proxy", pxyUrl)
 
@@ -186,6 +205,26 @@ func TestAppUnknownCommand(t *testing.T) {
 	assertLines(t, result.Stderr(), map[int]compareFunc{
 		0: equals(`ERROR "unknown-command": command not found`),
 	})
+}
+
+func TestAppHelp(t *testing.T) {
+	t.Parallel()
+
+	_, s5cmd := setup(t)
+
+	// without any specific command
+	cmd := s5cmd("--help")
+	result := icmd.RunCmd(cmd)
+
+	result.Assert(t, icmd.Success)
+
+	// with commands
+	for _, command := range command.Commands() {
+		cmd := s5cmd(command.Name, "--help")
+		result = icmd.RunCmd(cmd)
+
+		result.Assert(t, icmd.Success)
+	}
 }
 
 func TestUsageError(t *testing.T) {